% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{plot.pi}
\alias{plot.pi}
\title{Simple estimate of pi using a simulation}
\usage{
\method{plot}{pi}(x)
}
\arguments{
\item{x}{A \code{class} (pi) used to pass the points of a simulation into a graph}
}
\value{
A \code{plot} containing the circle,square and all the points
}
\description{
Plot an objet of class "pi" usually created by estimate_pi
}
\examples{
x <- estimate_pi( B=5000, seed = 10)

plot.pi(x)
}
\author{
Julien Kleinmann, Constance Havret, Lumia Claramunt, Sana Ghali,
Azza Akamoun
}
